resort-agentic-ai/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
streamlit
openai
langchain
langgraph
